+++ src/App.tsx
@@ -1,6 +1,6 @@
 import './styles/reset.css';
 import ProductsPage from './pages/ProductsPage/ProductsPage';
-import { ToastProvider } from './context/ToastProvider';
+import { ToastProvider } from './context/Toast/ToastProvider';
 
 function App() {
   return (
