+++ src/context/Toast/ToastContext.tsx
@@ -1,5 +1,5 @@
 import { createContext } from 'react';
-import { ToastvariantType } from '../components/common/Toast/Toast';
+import { ToastvariantType } from '../../components/common/Toast/Toast';
 
 export interface ToastContextType {
   showToast: (params: { text: string; variant: ToastvariantType }) => void;
